observability: extract shared otel types
