+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
     "crates/atm-agent-mcp",
     "crates/atm-tui",
     "crates/sc-observability",
+    "crates/sc-observability-types",
     "crates/sc-observability-otlp",
     "crates/sc-composer",
     "crates/sc-compose",
@@ -48,4 +49,5 @@ agent-team-mail-daemon-launch = { path = "crates/atm-daemon-launch", version = "
 agent-team-mail-daemon = { path = "crates/atm-daemon", version = "=0.46.1" }
 sc-composer = { path = "crates/sc-composer", version = "=0.46.1" }
 sc-observability = { path = "crates/sc-observability", version = "=0.46.1" }
+sc-observability-types = { path = "crates/sc-observability-types", version = "=0.46.1" }
 sc-observability-otlp = { path = "crates/sc-observability-otlp", version = "=0.46.1" }
