+++ crates/atm-daemon/Cargo.toml
@@ -30,6 +30,8 @@ agent-team-mail-ci-monitor.workspace = true
 agent-team-mail-core.workspace = true
 agent-team-mail-daemon-launch.workspace = true
 sc-observability.workspace = true
+sc-observability-otlp.workspace = true
+sc-observability-types.workspace = true
 thiserror.workspace = true
 anyhow.workspace = true
 serde.workspace = true
