+++ crates/atm-daemon/src/daemon/event_loop.rs
@@ -13,6 +13,9 @@ use crate::daemon::{
     },
     graceful_shutdown, spool_drain_loop, start_socket_server, watch_inboxes,
 };
+use crate::daemon::observability::{
+    export_metric_records_best_effort, export_trace_records_best_effort, otel_config_from_env,
+};
 use crate::plugin::{Capability, FailedPluginInit, PluginContext, PluginRegistry};
 use crate::plugins::worker_adapter::AgentState;
 use agent_team_mail_core::event_log::{EventFields, emit_event_best_effort};
@@ -20,10 +23,7 @@ use agent_team_mail_core::schema::TeamConfig;
 use agent_team_mail_core::team_config_store::TeamConfigStore;
 use anyhow::{Context, Result};
 use chrono::Utc;
-use sc_observability::{
-    MetricKind, MetricRecord, OtelConfig, TraceRecord, TraceStatus,
-    export_metric_records_best_effort, export_trace_records_best_effort,
-};
+use sc_observability_types::{MetricKind, MetricRecord, TraceRecord, TraceStatus};
 use serde_json::Value;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
@@ -629,7 +629,7 @@ pub async fn run(
                                 &dispatch_trace_id,
                                 "dispatch_message",
                             );
-                        let otel_config = OtelConfig::from_env();
+                        let otel_config = otel_config_from_env();
                         let mut dispatch_failed = false;
 
                         emit_event_best_effort(EventFields {
@@ -1684,7 +1684,7 @@ async fn status_writer_loop(
     let otel = build_otel_health(&ctx);
     export_metric_records_best_effort(
         &build_daemon_health_metric_records(&logging, &otel),
-        &OtelConfig::from_env(),
+        &otel_config_from_env(),
     );
     if let Err(e) = status_writer.write_daemon_touch(&teams) {
         error!("Failed to write daemon touch sidecar: {}", e);
@@ -1714,7 +1714,7 @@ async fn status_writer_loop(
                 let otel = build_otel_health(&ctx);
                 export_metric_records_best_effort(
                     &build_daemon_health_metric_records(&logging, &otel),
-                    &OtelConfig::from_env(),
+                    &otel_config_from_env(),
                 );
 
                 if let Err(e) = status_writer.write_status(plugin_statuses, teams, logging, otel) {
@@ -2010,7 +2010,7 @@ mod tests {
     use crate::plugins::worker_adapter::AgentState;
     use agent_team_mail_core::event_log::span_id_for_action;
     use agent_team_mail_core::schema::InboxMessage;
-    use sc_observability::TraceStatus;
+    use sc_observability_types::TraceStatus;
     use std::collections::HashMap;
     use std::fs as stdfs;
     use std::time::Duration;
