+++ crates/sc-observability-otlp/Cargo.toml
@@ -14,6 +14,7 @@ categories = ["development-tools", "network-programming"]
 [dependencies]
 chrono.workspace = true
 reqwest.workspace = true
+sc-observability-types.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
