+++ crates/sc-observability/Cargo.toml
@@ -15,6 +15,7 @@ categories = ["command-line-utilities", "development-tools"]
 agent-team-mail-core.workspace = true
 anyhow.workspace = true
 chrono = { version = "0.4", default-features = true }
+sc-observability-types.workspace = true
 sc-observability-otlp.workspace = true # approved direction: sc-observability -> otlp. Reverse (otlp -> sc-observability) is prohibited; see GH-876.
 serde.workspace = true
 serde_json.workspace = true
